refactor(workspace): split app-left shell helpers
